[Feat] DateDetail API 연결 (#86)

+++ DATEROAD-iOS/DATEROAD-iOS/Network/DateSchedule/DTO/GetDateDetailResponse.swift
@@ -26,7 +26,7 @@ struct GetDateDetailResponse: Codable {
 // MARK: - Place
 struct Place: Codable {
     let title: String
-    let duration: Double
+    let duration: Float
     let sequence: Int
 }
